+++ src/argilla/server/search_engine.py
@@ -141,9 +141,14 @@ class SearchEngine:
             for record in records
         ]
 
-        _, errors = await helpers.async_bulk(client=self.client, actions=bulk_actions, raise_on_error=False)
-        if errors:
-            raise RuntimeError(errors)
+        await self._bulk_op(bulk_actions)
+
+    async def delete_records(self, dataset: Dataset, records: Iterable[Record]):
+        index_name = await self._get_index_or_raise(dataset)
+
+        bulk_actions = [{"_op_type": "delete", "_id": record.id, "_index": index_name} for record in records]
+
+        await self._bulk_op(bulk_actions)
 
     async def update_record_response(self, response: Response):
         record = response.record
@@ -289,6 +294,11 @@ class SearchEngine:
         else:
             # See https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-terms-query.html
             return [{"term": {f"{user_response_field}.status": status_filter.status}}]
+
+    async def _bulk_op(self, actions: List[Dict[str, Any]]):
+        _, errors = await helpers.async_bulk(client=self.client, actions=actions, raise_on_error=False)
+        if errors:
+            raise RuntimeError(errors)
 
 
 async def get_search_engine() -> AsyncGenerator[SearchEngine, None]: