+++ tests/unit/server/api/v1/test_records.py
@@ -18,7 +18,7 @@ from uuid import UUID, uuid4
 
 import pytest
 from argilla._constants import API_KEY_HEADER_NAME
-from argilla.server.models import Response, Suggestion, User, UserRole
+from argilla.server.models import Record, Response, Suggestion, User, UserRole
 from argilla.server.search_engine import SearchEngine
 from fastapi.testclient import TestClient
 from sqlalchemy import func, select
@@ -40,6 +40,7 @@ from tests.factories import (
 
 if TYPE_CHECKING:
     from argilla.server.models import Dataset
+    from httpx import AsyncClient
     from sqlalchemy.ext.asyncio import AsyncSession
 
 
@@ -945,3 +946,57 @@ async def test_create_record_suggestion_as_annotator(async_client: "AsyncClient"
     )
 
     assert response.status_code == 403
+
+
+@pytest.mark.parametrize("role", [UserRole.owner, UserRole.admin])
+@pytest.mark.asyncio
+async def test_delete_record(
+    async_client: "AsyncClient", db: "AsyncSession", mock_search_engine: "SearchEngine", role: UserRole
+):
+    dataset = await DatasetFactory.create()
+    record = await RecordFactory.create(dataset=dataset)
+    user = await UserFactory.create(role=role, workspaces=[dataset.workspace])
+
+    response = await async_client.delete(f"/api/v1/records/{record.id}", headers={API_KEY_HEADER_NAME: user.api_key})
+
+    assert response.status_code == 200
+    assert response.json() == {
+        "id": str(record.id),
+        "fields": record.fields,
+        "metadata": None,
+        "external_id": record.external_id,
+        "inserted_at": record.inserted_at.isoformat(),
+        "updated_at": record.updated_at.isoformat(),
+    }
+    assert (await db.execute(select(func.count(Record.id)))).scalar() == 0
+    mock_search_engine.delete_records.assert_called_once_with(dataset=dataset, records=[record])
+
+
+@pytest.mark.asyncio
+async def test_delete_record_as_admin_from_another_workspace(async_client: "AsyncClient", db: "AsyncSession"):
+    dataset = await DatasetFactory.create()
+    record = await RecordFactory.create(dataset=dataset)
+    user = await UserFactory.create(role=UserRole.admin)
+
+    response = await async_client.delete(f"/api/v1/records/{record.id}", headers={API_KEY_HEADER_NAME: user.api_key})
+
+    assert response.status_code == 403
+    assert (await db.execute(select(func.count(Record.id)))).scalar() == 1
+
+
+@pytest.mark.asyncio
+async def test_delete_record_as_annotator(async_client: "AsyncClient"):
+    annotator = await UserFactory.create(role=UserRole.annotator)
+    record = await RecordFactory.create()
+
+    response = await async_client.delete(
+        f"/api/v1/records/{record.id}", headers={API_KEY_HEADER_NAME: annotator.api_key}
+    )
+
+    assert response.status_code == 403
+
+
+@pytest.mark.asyncio
+async def test_delete_record_non_existent(async_client: "AsyncClient", owner_auth_header: dict):
+    response = await async_client.delete(f"/api/v1/records/{uuid4()}", headers=owner_auth_header)
+    assert response.status_code == 404